update db models

+++ src/main/java/com/joaco/restaurantflowserver/api/MenuItemApi.java
@@ -31,7 +31,7 @@ public class MenuItemApi {
   public ResponseEntity<MenuItem> createMenuItem(@RequestBody MenuItemDto menuItem) {
 
     MenuItem save = new MenuItem(
-        null,
+        (Integer) null,
         menuItem.name(),
         menuItem.items(),
         menuItem.price(),
